Fix bug with recurring events and multiple calendars.
The `recurringEvents` array was not being reset on each new calendar.
This would cause "calendar contamination" when using this script with
more than one Google Calendar, as any recurring events processed on the
first calendar would also end up added to the second calendar.

Out of an abundance of caution, I've added a section of code in the main
loop that resets all "per-calendar" global variables. Longer-term, we
might think about how to factor the global variables out entirely.