Sync Tasks without dueDate and avoid error
This is my first commit ever and I barely know how to even do this and how JS works. I am using NC Tasks and want to sync them to Google, but many of them have no "dueDate", so my idea is to set one at today+30 days (or whatever) if there is none available (could be a var in the Code.gs, but Idk how to). Thank you.